Import decklist from MTG Arena and collect information about the individual cards
Magic: The Gathering is a 2+ player game where each player assembles a deck of 60 cards and uses them in a game against their opponent.

Magic cards are discrete objects with a variety of variables common across most of the cards. This program uses a txt file from a digital version of the card game and creates a local json database of the cards contained in the deck and their relevant information. Various sub-programs display information on the cards as well as simulate opening hands.